5068: the double vision GUI / POT viewport fix
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@8511 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/ai/EAS/EAS.cpp
@@ -267,7 +267,7 @@ void tdmEAS::SetupRoutesBetweenClusters()
 
 	//stgatilov #4755: precompute areas for clusters to improve asymptotic time complexity
 	std::vector<int> areaOfCluster(_clusterInfo.size());
-	for (int cluster = 0; cluster < _clusterInfo.size(); cluster++)
+	for (size_t cluster = 0; cluster < _clusterInfo.size(); cluster++)
 		areaOfCluster[cluster] = _aas->GetAreaInCluster(cluster);
 
 	for ( std::size_t startCluster = 0 ; startCluster < _clusterInfo.size() ; startCluster++ )
